handle invalid integer attribute values
and centralize some of the functions to access opportunistic attributes
to JobFunctions
